# model/model_access_type.go
/*
 * NSSF NS Selection
 *
 * NSSF Network Slice Selection Service
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package model

type AccessType string

// List of AccessType
const (
	_3_GPP_ACCESS AccessType = "3GPP_ACCESS"
	NON_3_GPP_ACCESS AccessType = "NON_3GPP_ACCESS"
)