# model/model_allowed_nssai.go
/*
 * NSSF NS Selection
 *
 * NSSF Network Slice Selection Service
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package model

type AllowedNssai struct {

	AllowedSnssaiList []AllowedSnssai `json:"allowedSnssaiList"`

	AccessType *AccessType `json:"accessType"`
}