# model/model_allowed_snssai.go
/*
 * NSSF NS Selection
 *
 * NSSF Network Slice Selection Service
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package model

type AllowedSnssai struct {

	AllowedSnssai *Snssai `json:"allowedSnssai"`

	NsiInformationList []NsiInformation `json:"nsiInformationList,omitempty"`

	MappedHomeSnssai *Snssai `json:"mappedHomeSnssai,omitempty"`
}