# model/model_authorized_network_slice_info.go
/*
 * NSSF NS Selection
 *
 * NSSF Network Slice Selection Service
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package model

type AuthorizedNetworkSliceInfo struct {

	AllowedNssaiList []AllowedNssai `json:"allowedNssaiList,omitempty"`

	ConfiguredNssai []ConfiguredSnssai `json:"configuredNssai,omitempty"`

	TargetAmfSet string `json:"targetAmfSet,omitempty"`

	CandidateAmfList []string `json:"candidateAmfList,omitempty"`

	RejectedNssaiInPlmn []Snssai `json:"rejectedNssaiInPlmn,omitempty"`

	RejectedNssaiInTa []Snssai `json:"rejectedNssaiInTa,omitempty"`

	NsiInformation *NsiInformation `json:"nsiInformation,omitempty"`

	SupportedFeatures string `json:"supportedFeatures,omitempty"`

	NrfAmfSet string `json:"nrfAmfSet,omitempty"`
}