# model/model_configured_snssai.go
/*
 * NSSF NS Selection
 *
 * NSSF Network Slice Selection Service
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package model

type ConfiguredSnssai struct {

	ConfiguredSnssai *Snssai `json:"configuredSnssai"`

	MappedHomeSnssai *Snssai `json:"mappedHomeSnssai,omitempty"`
}