# model/model_invalid_param.go
/*
 * NSSF NS Selection
 *
 * NSSF Network Slice Selection Service
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package model

type InvalidParam struct {

	Param string `json:"param"`

	Reason string `json:"reason,omitempty"`
}