# model/model_mapping_of_snssai.go
/*
 * NSSF NS Selection
 *
 * NSSF Network Slice Selection Service
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package model

type MappingOfSnssai struct {

	ServingSnssai *Snssai `json:"servingSnssai"`

	HomeSnssai *Snssai `json:"homeSnssai"`
}