# model/model_nsi_information.go
/*
 * NSSF NS Selection
 *
 * NSSF Network Slice Selection Service
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package model

type NsiInformation struct {

	NrfId string `json:"nrfId"`

	NsiId string `json:"nsiId,omitempty"`
}