# model/model_plmn_id.go
/*
 * NSSF NS Selection
 *
 * NSSF Network Slice Selection Service
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package model

type PlmnId struct {

	Mcc string `json:"mcc"`

	Mnc string `json:"mnc"`
}