# model/model_slice_info_for_registration.go
/*
 * NSSF NS Selection
 *
 * NSSF Network Slice Selection Service
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package model

type SliceInfoForRegistration struct {

	SubscribedNssai []SubscribedSnssai `json:"subscribedNssai,omitempty"`

	AllowedNssaiCurrentAccess *AllowedNssai `json:"allowedNssaiCurrentAccess,omitempty"`

	AllowedNssaiOtherAccess *AllowedNssai `json:"allowedNssaiOtherAccess,omitempty"`

	SNssaiForMapping []Snssai `json:"sNssaiForMapping,omitempty"`

	RequestedNssai []Snssai `json:"requestedNssai,omitempty"`

	DefaultConfiguredSnssaiInd bool `json:"defaultConfiguredSnssaiInd,omitempty"`

	MappingOfNssai []MappingOfSnssai `json:"mappingOfNssai,omitempty"`

	RequestMapping bool `json:"requestMapping,omitempty"`
}