# model/model_snssai.go
/*
 * NSSF NS Selection
 *
 * NSSF Network Slice Selection Service
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package model

type Snssai struct {

	Sst int32 `json:"sst"`

	Sd string `json:"sd,omitempty"`
}