# model/model_subscribed_snssai.go
/*
 * NSSF NS Selection
 *
 * NSSF Network Slice Selection Service
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package model

type SubscribedSnssai struct {

	SubscribedSnssai *Snssai `json:"subscribedSnssai"`

	DefaultIndication bool `json:"defaultIndication,omitempty"`
}