# model/model_tai.go
/*
 * NSSF NS Selection
 *
 * NSSF Network Slice Selection Service
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package model

type Tai struct {

	PlmnId *PlmnId `json:"plmnId"`

	Tac string `json:"tac"`
}